requirements.txt
Django==1.8.6
django-admin-bootstrapped==2.5.7
django-allauth==0.24.1
django-bootstrap3-datetimepicker==2.2.3
gunicorn==19.3.0
oauthlib==1.0.3
python-openid==2.2.5
pytz==2015.7
requests==2.9.1
requests-oauthlib==0.6.0
